Add film controller tests. Move test to directory. Import from index
